+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+* Add `Lang::is_machine_translation`.
+    - Adds support for `machine` variant, as in `en-machine` to indicate a resource that was machine translated.
+    - `LangMap` gives priority for human translations of different regions over machine translations.
+
 * Improve `cargo zng l10n --check`.
     - Check the Fluent syntax of localizations.
     - Check if localized messages match the template.